Elsys - HomeWork 16/20
16) extend the clickhandler that has been defined in 15) as follows: ask
the user for confirmation (see confirm() function) and do nothing if the
user declines confirmation. If the user accepts the confirmation - send
a DELETE request to http://jsonplaceholder.typicode.com/posts/ and if
the request succeeds - remove the element created for this post.

+++ ELSYS/WebContent/scripts/elsys.js
@@ -22,15 +22,16 @@ $(document).ready(function() {
 			$("#posts").append("<li>"+ response[i].body +"</li>")
 			.append("<button id=\ " + response[i].id + " class=delete" + ">Delete</button>");
 		};
-	}).then(function(){
 		$(".delete").click(function() {
-			var idButton = $(this).attr("id");
-			$.ajax(ENDPOINT + "/" + idButton, {
-				method: "DELETE",
-				dataType: "json"
-			}).then(function(response){
-				alert("deleting");
-			});
+			if(confirm("Deleting "+response[i].body+"!")){
+				var idButton = $(this).attr("id");
+				$.ajax(ENDPOINT + "/" + idButton, {
+					method: "DELETE",
+					dataType: "json"
+				}).then(function(response){
+					alert("Deleted!");
+				});		
+			};
 		});
 	});
